Added camera movement, fixed BVH bug connected to box intersections
Created a small system to deal with the input from the template.
Box intersections now also happen if the point is inside the box, in which case the distance to the intersection point is negative. Probably wrong behaviour during BVH ray tracing if the origin of the ray is inside both bounding boxes of the branch.
Changed the camera from orthographic to perspective, swapped around world space axes to make it more intuitive.